pyproject.toml
[project]
name = "rag-study"
version = "0.1.0"
description = ""
authors = [
    {name = "Ssoon-m",email = "tkksm2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    # 유틸리티 관련 패키지
    "python-dotenv (>=1.1.0,<2.0.0)",
    # 언어 및 자연어 처리 관련 패키지
    "langchain-openai (>=0.3.16,<0.4.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"